requirements.txt
flask
joblib
pandas
scikit-learn
gunicorn
imbalanced-learn
python-dotenv
pytest
pytest-mock